Project added
In this project, I tried to use a list of employees that I had created manually and perform CRUD Operation on it.
Currently facing an issue in this project that when I try to perform any CRUD operation on this list, my constructor creates and object of the list again and again.
For this, I need to store the list and then perform the operation on it.